feat: Add feature gates for CLN, LND, fakewallet and LNbits backends
