+++ .github/workflows/ci.yml
@@ -105,11 +105,19 @@ jobs:
             -p cdk-fake-wallet,
             --bin cdk-cli,
             --bin cdk-mintd,
-            --bin cdk-mintd --no-default-features --features swagger,
-            --bin cdk-mintd --no-default-features --features redis,
-            --bin cdk-mintd --no-default-features --features "redis swagger",
-            --bin cdk-mintd --no-default-features --features management-rpc,
+            --bin cdk-mintd --features redis,
+            --bin cdk-mintd --features "redis swagger redb",
             --bin cdk-mintd --no-default-features --features redb,
+            --bin cdk-mintd --no-default-features --features lnd,
+            --bin cdk-mintd --no-default-features --features cln,
+            --bin cdk-mintd --no-default-features --features lnbits,
+            --bin cdk-mintd --no-default-features --features fakewallet,
+            --bin cdk-mintd --no-default-features --features "management-rpc lnd",
+            --bin cdk-mintd --no-default-features --features "management-rpc cln",
+            --bin cdk-mintd --no-default-features --features "management-rpc lnbits",
+            --bin cdk-mintd --no-default-features --features "swagger lnd",
+            --bin cdk-mintd --no-default-features --features "swagger cln",
+            --bin cdk-mintd --no-default-features --features "swagger lnbits",
             --bin cdk-mint-cli,
           ]
     steps:
