+++ crates/cdk-mintd/Cargo.toml
@@ -10,12 +10,17 @@ description = "CDK mint binary"
 rust-version = "1.75.0"
 
 [features]
-default = ["management-rpc"]
+default = ["management-rpc", "cln", "lnd", "lnbits", "fakewallet"]
+# Ensure at least one lightning backend is enabled
 swagger = ["cdk-axum/swagger", "dep:utoipa", "dep:utoipa-swagger-ui"]
 redis = ["cdk-axum/redis"]
 management-rpc = ["cdk-mint-rpc"]
 # MSRV is not commited to with redb enabled
 redb = ["dep:cdk-redb"]
+cln = ["dep:cdk-cln"]
+lnd = ["dep:cdk-lnd"]
+lnbits = ["dep:cdk-lnbits"]
+fakewallet = ["dep:cdk-fake-wallet"]
 
 [dependencies]
 anyhow.workspace = true
@@ -30,10 +35,10 @@ cdk-redb = { workspace = true, features = [
 cdk-sqlite = { workspace = true, features = [
     "mint",
 ] }
-cdk-cln.workspace = true
-cdk-lnbits.workspace = true
-cdk-lnd.workspace = true
-cdk-fake-wallet.workspace = true
+cdk-cln = { workspace = true, optional = true }
+cdk-lnbits = { workspace = true, optional = true }
+cdk-lnd = { workspace = true, optional = true }
+cdk-fake-wallet = { workspace = true, optional = true }
 cdk-axum.workspace = true
 cdk-mint-rpc = { workspace = true, optional = true }
 config = { version = "0.13.3", features = ["toml"] }
