+++ crates/cdk-mintd/src/main.rs
@@ -14,6 +14,13 @@ use axum::Router;
 use bip39::Mnemonic;
 use cdk::cdk_database::{self, MintDatabase};
 use cdk::mint::{MintBuilder, MintMeltLimits};
+// Feature-gated imports
+#[cfg(any(
+    feature = "cln",
+    feature = "lnbits",
+    feature = "lnd",
+    feature = "fakewallet"
+))]
 use cdk::nuts::nut17::SupportedMethods;
 use cdk::nuts::nut19::{CachedEndpoint, Method as NUT19Method, Path as NUT19Path};
 use cdk::nuts::{ContactInfo, CurrencyUnit, MintVersion, PaymentMethod};
@@ -39,6 +46,17 @@ use tracing_subscriber::EnvFilter;
 use utoipa::OpenApi;
 
 const CARGO_PKG_VERSION: Option<&'static str> = option_env!("CARGO_PKG_VERSION");
+
+// Ensure at least one lightning backend is enabled at compile time
+#[cfg(not(any(
+    feature = "cln",
+    feature = "lnbits",
+    feature = "lnd",
+    feature = "fakewallet"
+)))]
+compile_error!(
+    "At least one lightning backend feature must be enabled: cln, lnbits, lnd, or fakewallet"
+);
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
@@ -149,6 +167,7 @@ async fn main() -> anyhow::Result<()> {
     };
 
     match settings.ln.ln_backend {
+        #[cfg(feature = "cln")]
         LnBackend::Cln => {
             let cln_settings = settings
                 .cln
@@ -171,6 +190,7 @@ async fn main() -> anyhow::Result<()> {
 
             mint_builder = mint_builder.add_supported_websockets(nut17_supported);
         }
+        #[cfg(feature = "lnbits")]
         LnBackend::LNbits => {
             let lnbits_settings = settings.clone().lnbits.expect("Checked on config load");
             let lnbits = lnbits_settings
@@ -187,6 +207,7 @@ async fn main() -> anyhow::Result<()> {
 
             mint_builder = mint_builder.add_supported_websockets(nut17_supported);
         }
+        #[cfg(feature = "lnd")]
         LnBackend::Lnd => {
             let lnd_settings = settings.clone().lnd.expect("Checked at config load");
             let lnd = lnd_settings
@@ -204,6 +225,7 @@ async fn main() -> anyhow::Result<()> {
 
             mint_builder = mint_builder.add_supported_websockets(nut17_supported);
         }
+        #[cfg(feature = "fakewallet")]
         LnBackend::FakeWallet => {
             let fake_wallet = settings.clone().fake_wallet.expect("Fake wallet defined");
